Create README.md
In this game, Princess Peach and Bowser randomly appear on a 3x3 grid of tiles. Your goal is to click on as many Princess tiles as possible to increase your score, but be careful not to click on Bowser tiles, as doing so will end the game. The game features multiple Princess and Bowser tiles appearing simultaneously, adding to the challenge and excitement.